pyproject.toml
[tool.poetry]
name = "dspy-starter"
version = "0.1.0"
description = ""
authors = ["Anwesh43 <anweshthecool0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.16"
dspy = "^2.6.24"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
